refactor: migrate jest-config jest-fail-on-console config to jest-config-react
