+++ @ornikar/jest-config-react/test-setup-after-env.js
@@ -4,6 +4,13 @@
 
 const failOnConsole = require('jest-fail-on-console');
 
+const deprecatedReactLifeCycleMethods = [
+  'componentWillMount',
+  'componentWillUnmount',
+  'componentWillUpdate',
+  'componentWillReceiveProps',
+];
+
 failOnConsole({
   silenceMessage: (message) => {
     // setNativeProps is used by @react-spring, so we can't do anything about it
@@ -15,6 +22,21 @@ failOnConsole({
     if (
       message.includes(
         'componentWillMount has been renamed, and is not recommended for use. See https://reactjs.org/link/unsafe-component-lifecycles for detail',
+      )
+    ) {
+      return true;
+    }
+
+    // Do not ignore the act / await warning anymore once we are on React 18
+    if (message.includes('You called act(async () => ...) without await')) return true;
+
+    // Warning from @react-aria/ssr
+    if (message.startsWith('In React 18, SSRProvider is not necessary')) return true;
+
+    // Deprecated lifecycle methods are forbidden in our eslint config. This means any related warning is due to a dependency and can be ignored in tests.
+    if (
+      deprecatedReactLifeCycleMethods.some((lifecycleMethod) =>
+        message.includes(`${lifecycleMethod} has been renamed, and is not recommended for use.`),
       )
     ) {
       return true;
