+++ @ornikar/jest-config/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "chalk": "^4.0.0",
-    "jest-fail-on-console": "^3.1.1",
     "mockdate": "^3.0.2"
   },
   "peerDependencies": {
